# Gin_framework/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API สำหรับจัดการข้อมูล Albums",
        "title": "Album API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/albums": {
            "get": {
                "description": "ดึงรายการ albums ทั้งหมด",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get all albums",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Album"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "เพิ่ม album ใหม่",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Create a new album",
                "parameters": [
                    {
                        "description": "Album data",
                        "name": "album",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/albums/{id}": {
            "get": {
                "description": "ดึง album ตาม ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "Get album by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Album ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Album"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Album": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string",
                    "example": "John Coltrane"
                },
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "price": {
                    "type": "number",
                    "example": 56.99
                },
                "title": {
                    "type": "string",
                    "example": "Blue Train"
                }
            }
        }
    }
}